refactor(api): migrate agent endpoints to workspace and remove deprecated agent file management
