+++ console/src/api/modules/agent.ts
@@ -26,26 +26,29 @@ export const agentApi = {
     }),
 
   getAgentRunningConfig: () =>
-    request<AgentsRunningConfig>("/agent/running-config"),
+    request<AgentsRunningConfig>("/workspace/running-config"),
 
   updateAgentRunningConfig: (config: AgentsRunningConfig) =>
-    request<AgentsRunningConfig>("/agent/running-config", {
+    request<AgentsRunningConfig>("/workspace/running-config", {
       method: "PUT",
       body: JSON.stringify(config),
     }),
 
-  getAgentLanguage: () => request<{ language: string }>("/agent/language"),
+  getAgentLanguage: () => request<{ language: string }>("/workspace/language"),
 
   updateAgentLanguage: (language: string) =>
-    request<{ language: string; copied_files: string[] }>("/agent/language", {
-      method: "PUT",
-      body: JSON.stringify({ language }),
-    }),
+    request<{ language: string; copied_files: string[] }>(
+      "/workspace/language",
+      {
+        method: "PUT",
+        body: JSON.stringify({ language }),
+      },
+    ),
 
-  getAudioMode: () => request<{ audio_mode: string }>("/agent/audio-mode"),
+  getAudioMode: () => request<{ audio_mode: string }>("/workspace/audio-mode"),
 
   updateAudioMode: (audio_mode: string) =>
-    request<{ audio_mode: string }>("/agent/audio-mode", {
+    request<{ audio_mode: string }>("/workspace/audio-mode", {
       method: "PUT",
       body: JSON.stringify({ audio_mode }),
     }),
@@ -54,22 +57,22 @@ export const agentApi = {
     request<{
       providers: { id: string; name: string; available: boolean }[];
       configured_provider_id: string;
-    }>("/agent/transcription-providers"),
+    }>("/workspace/transcription-providers"),
 
   updateTranscriptionProvider: (provider_id: string) =>
-    request<{ provider_id: string }>("/agent/transcription-provider", {
+    request<{ provider_id: string }>("/workspace/transcription-provider", {
       method: "PUT",
       body: JSON.stringify({ provider_id }),
     }),
 
   getTranscriptionProviderType: () =>
     request<{ transcription_provider_type: string }>(
-      "/agent/transcription-provider-type",
+      "/workspace/transcription-provider-type",
     ),
 
   updateTranscriptionProviderType: (transcription_provider_type: string) =>
     request<{ transcription_provider_type: string }>(
-      "/agent/transcription-provider-type",
+      "/workspace/transcription-provider-type",
       {
         method: "PUT",
         body: JSON.stringify({ transcription_provider_type }),
@@ -81,5 +84,5 @@ export const agentApi = {
       available: boolean;
       ffmpeg_installed: boolean;
       whisper_installed: boolean;
-    }>("/agent/local-whisper-status"),
+    }>("/workspace/local-whisper-status"),
 };
