+++ console/src/api/modules/agents.ts
@@ -6,7 +6,6 @@ import type {
   AgentProfileRef,
   ReorderAgentsResponse,
 } from "../types/agents";
-import type { MdFileInfo, MdFileContent } from "../types/workspace";
 
 // Multi-agent management API
 export const agentsApi = {
@@ -53,26 +52,4 @@ export const agentsApi = {
         body: JSON.stringify({ enabled }),
       },
     ),
-
-  // Agent workspace files
-  listAgentFiles: (agentId: string) =>
-    request<MdFileInfo[]>(`/agents/${agentId}/files`),
-
-  readAgentFile: (agentId: string, filename: string) =>
-    request<MdFileContent>(
-      `/agents/${agentId}/files/${encodeURIComponent(filename)}`,
-    ),
-
-  writeAgentFile: (agentId: string, filename: string, content: string) =>
-    request<{ written: boolean; filename: string }>(
-      `/agents/${agentId}/files/${encodeURIComponent(filename)}`,
-      {
-        method: "PUT",
-        body: JSON.stringify({ content }),
-      },
-    ),
-
-  // Agent memory files
-  listAgentMemory: (agentId: string) =>
-    request<MdFileInfo[]>(`/agents/${agentId}/memory`),
 };
