+++ console/src/api/modules/workspace.ts
@@ -38,7 +38,7 @@ export interface WorkspaceDownloadResult {
 
 export const workspaceApi = {
   listFiles: () =>
-    request<MdFileInfo[]>("/agent/files").then((files) =>
+    request<MdFileInfo[]>("/workspace/files").then((files) =>
       files.map((file) => ({
         ...file,
         updated_at: new Date(file.modified_time).getTime(),
@@ -46,11 +46,11 @@ export const workspaceApi = {
     ),
 
   loadFile: (fileName: string) =>
-    request<MdFileContent>(`/agent/files/${encodeURIComponent(fileName)}`),
+    request<MdFileContent>(`/workspace/files/${encodeURIComponent(fileName)}`),
 
   saveFile: (fileName: string, content: string) =>
     request<Record<string, unknown>>(
-      `/agent/files/${encodeURIComponent(fileName)}`,
+      `/workspace/files/${encodeURIComponent(fileName)}`,
       {
         method: "PUT",
         body: JSON.stringify({ content }),
@@ -114,7 +114,7 @@ export const workspaceApi = {
   },
 
   listDailyMemory: () =>
-    request<MdFileInfo[]>("/agent/memory").then((files) =>
+    request<MdFileInfo[]>("/workspace/memory").then((files) =>
       files.map((file) => {
         const date = file.filename.replace(".md", "");
         return {
@@ -126,11 +126,11 @@ export const workspaceApi = {
     ),
 
   loadDailyMemory: (date: string) =>
-    request<MdFileContent>(`/agent/memory/${encodeURIComponent(date)}.md`),
+    request<MdFileContent>(`/workspace/memory/${encodeURIComponent(date)}.md`),
 
   saveDailyMemory: (date: string, content: string) =>
     request<Record<string, unknown>>(
-      `/agent/memory/${encodeURIComponent(date)}.md`,
+      `/workspace/memory/${encodeURIComponent(date)}.md`,
       {
         method: "PUT",
         body: JSON.stringify({ content }),
@@ -138,10 +138,11 @@ export const workspaceApi = {
     ),
 
   // System prompt files management
-  getSystemPromptFiles: () => request<string[]>("/agent/system-prompt-files"),
+  getSystemPromptFiles: () =>
+    request<string[]>("/workspace/system-prompt-files"),
 
   setSystemPromptFiles: (files: string[]) =>
-    request<string[]>("/agent/system-prompt-files", {
+    request<string[]>("/workspace/system-prompt-files", {
       method: "PUT",
       body: JSON.stringify(files),
     }),
