+++ src/qwenpaw/app/routers/agent_scoped.py
@@ -56,7 +56,6 @@ def create_agent_scoped_router() -> APIRouter:
     Returns:
         APIRouter with all sub-routers mounted under /{agentId}/
     """
-    from .agent import router as agent_router
     from .skills import router as skills_router
     from .tools import router as tools_router
     from .config import router as config_router
@@ -68,11 +67,9 @@ def create_agent_scoped_router() -> APIRouter:
     from .plugins import router as plugins_router
     from .plan import router as plan_router
 
-    # Create parent router with agentId parameter
     router = APIRouter(prefix="/agents/{agentId}", tags=["agent-scoped"])
 
     # Include all agent-specific sub-routers (they keep their own prefixes)
-    # /agents/{agentId}/agent/* -> agent_router
     # /agents/{agentId}/chats/* -> chats_router
     # /agents/{agentId}/config/* -> config_router (channels, heartbeat)
     # /agents/{agentId}/cron/* -> cron_router
@@ -80,7 +77,6 @@ def create_agent_scoped_router() -> APIRouter:
     # /agents/{agentId}/skills/* -> skills_router
     # /agents/{agentId}/tools/* -> tools_router
     # /agents/{agentId}/workspace/* -> workspace_router
-    router.include_router(agent_router)
     router.include_router(chats_router)
     router.include_router(config_router)
     router.include_router(cron_router)
