+++ src/qwenpaw/app/routers/workspace.py
@@ -1,5 +1,9 @@
 # -*- coding: utf-8 -*-
-"""Workspace API – download / upload the entire WORKING_DIR as a zip."""
+"""Workspace API – download / upload the entire WORKING_DIR as a zip.
+
+Also includes agent file management, language settings, audio/transcription
+configuration, running config, and system prompt files.
+"""
 
 from __future__ import annotations
 
@@ -11,11 +15,41 @@ import zipfile
 from datetime import datetime, timezone
 from pathlib import Path
 
-from fastapi import APIRouter, HTTPException, UploadFile, File, Request
+from fastapi import APIRouter, Body, HTTPException, UploadFile, File, Request
 from fastapi.responses import StreamingResponse
+from pydantic import BaseModel, Field
+
+from ..utils import schedule_agent_reload
+from ...config import (
+    load_config,
+    save_config,
+    AgentsRunningConfig,
+)
+from ...config.config import load_agent_config, save_agent_config
+from ...agents.memory.agent_md_manager import AgentMdManager
+from ...agents.templates import get_workspace_md_template_id
+from ...agents.utils import copy_workspace_md_files
+from ...constant import BUILTIN_QA_AGENT_ID
+from ..agent_context import get_agent_for_request
 
 
 router = APIRouter(prefix="/workspace", tags=["workspace"])
+
+
+class MdFileInfo(BaseModel):
+    """Markdown file metadata."""
+
+    filename: str = Field(..., description="File name")
+    path: str = Field(..., description="File path")
+    size: int = Field(..., description="Size in bytes")
+    created_time: str = Field(..., description="Created time")
+    modified_time: str = Field(..., description="Modified time")
+
+
+class MdFileContent(BaseModel):
+    """Markdown file content."""
+
+    content: str = Field(..., description="File content")
 
 
 def _dir_stats(root: Path) -> tuple[int, int]:
@@ -46,6 +80,474 @@ def _zip_directory(root: Path) -> io.BytesIO:
                 zf.write(entry, arcname + "/")
     buf.seek(0)
     return buf
+
+
+# ---------------------------------------------------------------------------
+# Agent File Management Endpoints
+# ---------------------------------------------------------------------------
+
+
+@router.get(
+    "/files",
+    response_model=list[MdFileInfo],
+    summary="List working files",
+    description="List all working files (uses active agent)",
+)
+async def list_working_files(
+    request: Request,
+) -> list[MdFileInfo]:
+    """List working directory markdown files."""
+    try:
+        workspace = await get_agent_for_request(request)
+        workspace_manager = AgentMdManager(
+            str(workspace.workspace_dir),
+            agent_id=workspace.agent_id,
+        )
+        files = [
+            MdFileInfo.model_validate(file)
+            for file in workspace_manager.list_working_mds()
+        ]
+        return files
+    except Exception as exc:
+        raise HTTPException(status_code=500, detail=str(exc)) from exc
+
+
+@router.get(
+    "/files/{md_name}",
+    response_model=MdFileContent,
+    summary="Read a working file",
+    description="Read a working markdown file (uses active agent)",
+)
+async def read_working_file(
+    md_name: str,
+    request: Request,
+) -> MdFileContent:
+    """Read a working directory markdown file."""
+    try:
+        workspace = await get_agent_for_request(request)
+        workspace_manager = AgentMdManager(
+            str(workspace.workspace_dir),
+            agent_id=workspace.agent_id,
+        )
+        content = workspace_manager.read_working_md(md_name)
+        return MdFileContent(content=content)
+    except FileNotFoundError as exc:
+        raise HTTPException(status_code=404, detail=str(exc)) from exc
+    except Exception as exc:
+        raise HTTPException(status_code=500, detail=str(exc)) from exc
+
+
+@router.put(
+    "/files/{md_name}",
+    response_model=dict,
+    summary="Write a working file",
+    description="Create or update a working file (uses active agent)",
+)
+async def write_working_file(
+    md_name: str,
+    body: MdFileContent,
+    request: Request,
+) -> dict:
+    """Write a working directory markdown file."""
+    try:
+        workspace = await get_agent_for_request(request)
+        workspace_manager = AgentMdManager(
+            str(workspace.workspace_dir),
+            agent_id=workspace.agent_id,
+        )
+        workspace_manager.write_working_md(md_name, body.content)
+        return {"written": True}
+    except Exception as exc:
+        raise HTTPException(status_code=500, detail=str(exc)) from exc
+
+
+@router.get(
+    "/memory",
+    response_model=list[MdFileInfo],
+    summary="List memory files",
+    description="List all memory files (uses active agent)",
+)
+async def list_memory_files(
+    request: Request,
+) -> list[MdFileInfo]:
+    """List memory directory markdown files."""
+    try:
+        workspace = await get_agent_for_request(request)
+        workspace_manager = AgentMdManager(
+            str(workspace.workspace_dir),
+            agent_id=workspace.agent_id,
+        )
+        files = [
+            MdFileInfo.model_validate(file)
+            for file in workspace_manager.list_memory_mds()
+        ]
+        return files
+    except Exception as exc:
+        raise HTTPException(status_code=500, detail=str(exc)) from exc
+
+
+@router.get(
+    "/memory/{md_name}",
+    response_model=MdFileContent,
+    summary="Read a memory file",
+    description="Read a memory markdown file (uses active agent)",
+)
+async def read_memory_file(
+    md_name: str,
+    request: Request,
+) -> MdFileContent:
+    """Read a memory directory markdown file."""
+    try:
+        workspace = await get_agent_for_request(request)
+        workspace_manager = AgentMdManager(
+            str(workspace.workspace_dir),
+            agent_id=workspace.agent_id,
+        )
+        content = workspace_manager.read_memory_md(md_name)
+        return MdFileContent(content=content)
+    except FileNotFoundError as exc:
+        raise HTTPException(status_code=404, detail=str(exc)) from exc
+    except Exception as exc:
+        raise HTTPException(status_code=500, detail=str(exc)) from exc
+
+
+@router.put(
+    "/memory/{md_name}",
+    response_model=dict,
+    summary="Write a memory file",
+    description="Create or update a memory file (uses active agent)",
+)
+async def write_memory_file(
+    md_name: str,
+    body: MdFileContent,
+    request: Request,
+) -> dict:
+    """Write a memory directory markdown file."""
+    try:
+        workspace = await get_agent_for_request(request)
+        workspace_manager = AgentMdManager(
+            str(workspace.workspace_dir),
+            agent_id=workspace.agent_id,
+        )
+        workspace_manager.write_memory_md(md_name, body.content)
+        return {"written": True}
+    except Exception as exc:
+        raise HTTPException(status_code=500, detail=str(exc)) from exc
+
+
+@router.get(
+    "/language",
+    summary="Get agent language",
+    description="Get the language setting for agent MD files (en/zh/ru)",
+)
+async def get_agent_language(request: Request) -> dict:
+    """Get agent language setting for current agent."""
+    workspace = await get_agent_for_request(request)
+    agent_config = load_agent_config(workspace.agent_id)
+    return {
+        "language": agent_config.language,
+        "agent_id": workspace.agent_id,
+    }
+
+
+@router.put(
+    "/language",
+    summary="Update agent language",
+    description=(
+        "Update the language for agent MD files (en/zh/ru). "
+        "Optionally copies MD files for the new language to agent workspace."
+    ),
+)
+async def put_agent_language(
+    request: Request,
+    body: dict = Body(
+        ...,
+        description='Language setting, e.g. {"language": "zh"}',
+    ),
+) -> dict:
+    """
+    Update agent language and optionally re-copy MD files to agent workspace.
+    """
+    language = (body.get("language") or "").strip().lower()
+    valid = {"zh", "en", "ru"}
+    if language not in valid:
+        raise HTTPException(
+            status_code=400,
+            detail=(
+                f"Invalid language '{language}'. "
+                f"Must be one of: {', '.join(sorted(valid))}"
+            ),
+        )
+
+    workspace = await get_agent_for_request(request)
+    agent_id = workspace.agent_id
+
+    agent_config = load_agent_config(agent_id)
+    old_language = agent_config.language
+
+    agent_config.language = language
+    save_agent_config(agent_id, agent_config)
+
+    copied_files: list[str] = []
+    if old_language != language:
+        copied_files = copy_workspace_md_files(
+            language,
+            workspace.workspace_dir,
+            md_template_id=get_workspace_md_template_id(
+                agent_config.template_id
+                or ("qa" if agent_id == BUILTIN_QA_AGENT_ID else None),
+            ),
+            only_if_missing=False,
+        )
+
+    return {
+        "language": language,
+        "copied_files": copied_files,
+        "agent_id": agent_id,
+    }
+
+
+@router.get(
+    "/audio-mode",
+    summary="Get audio mode",
+    description=(
+        "Get the audio handling mode for incoming voice messages. "
+        'Values: "auto", "native".'
+    ),
+)
+async def get_audio_mode() -> dict:
+    """Get audio mode setting."""
+    config = load_config()
+    return {"audio_mode": config.agents.audio_mode}
+
+
+@router.put(
+    "/audio-mode",
+    summary="Update audio mode",
+    description=(
+        "Update how incoming audio/voice messages are handled. "
+        '"auto": transcribe if provider available, else file placeholder; '
+        '"native": send audio directly to model (may need ffmpeg).'
+    ),
+)
+async def put_audio_mode(
+    body: dict = Body(
+        ...,
+        description='Audio mode, e.g. {"audio_mode": "auto"}',
+    ),
+) -> dict:
+    """Update audio mode setting."""
+    raw = body.get("audio_mode")
+    audio_mode = (str(raw) if raw is not None else "").strip().lower()
+    valid = {"auto", "native"}
+    if audio_mode not in valid:
+        raise HTTPException(
+            status_code=400,
+            detail=(
+                f"Invalid audio_mode '{audio_mode}'. "
+                f"Must be one of: {', '.join(sorted(valid))}"
+            ),
+        )
+    config = load_config()
+    config.agents.audio_mode = audio_mode
+    save_config(config)
+    return {"audio_mode": audio_mode}
+
+
+@router.get(
+    "/transcription-provider-type",
+    summary="Get transcription provider type",
+    description=(
+        "Get the transcription provider type. "
+        'Values: "disabled", "whisper_api", "local_whisper".'
+    ),
+)
+async def get_transcription_provider_type() -> dict:
+    """Get transcription provider type setting."""
+    config = load_config()
+    return {
+        "transcription_provider_type": (
+            config.agents.transcription_provider_type
+        ),
+    }
+
+
+@router.put(
+    "/transcription-provider-type",
+    summary="Set transcription provider type",
+    description=(
+        "Set the transcription provider type. "
+        '"disabled": no transcription; '
+        '"whisper_api": remote Whisper endpoint; '
+        '"local_whisper": locally installed openai-whisper.'
+    ),
+)
+async def put_transcription_provider_type(
+    body: dict = Body(
+        ...,
+        description=(
+            "Provider type, e.g. "
+            '{"transcription_provider_type": "whisper_api"}'
+        ),
+    ),
+) -> dict:
+    """Set the transcription provider type."""
+    raw = body.get("transcription_provider_type")
+    provider_type = (str(raw) if raw is not None else "").strip().lower()
+    valid = {"disabled", "whisper_api", "local_whisper"}
+    if provider_type not in valid:
+        raise HTTPException(
+            status_code=400,
+            detail=(
+                f"Invalid transcription_provider_type '{provider_type}'. "
+                f"Must be one of: {', '.join(sorted(valid))}"
+            ),
+        )
+    config = load_config()
+    config.agents.transcription_provider_type = provider_type
+    save_config(config)
+    return {"transcription_provider_type": provider_type}
+
+
+@router.get(
+    "/local-whisper-status",
+    summary="Check local whisper availability",
+    description=(
+        "Check whether the local whisper provider can be used. "
+        "Returns availability of ffmpeg and openai-whisper."
+    ),
+)
+async def get_local_whisper_status() -> dict:
+    """Check local whisper dependencies."""
+    from ...agents.utils.audio_transcription import (
+        check_local_whisper_available,
+    )
+
+    return check_local_whisper_available()
+
+
+@router.get(
+    "/transcription-providers",
+    summary="List transcription providers",
+    description=(
+        "List providers capable of audio transcription (Whisper API). "
+        "Returns available providers and the configured selection."
+    ),
+)
+async def get_transcription_providers() -> dict:
+    """List transcription-capable providers and configured selection."""
+    from ...agents.utils.audio_transcription import (
+        get_configured_transcription_provider_id,
+        list_transcription_providers,
+    )
+
+    return {
+        "providers": list_transcription_providers(),
+        "configured_provider_id": (get_configured_transcription_provider_id()),
+    }
+
+
+@router.put(
+    "/transcription-provider",
+    summary="Set transcription provider",
+    description=(
+        "Set the provider to use for audio transcription. "
+        'Use empty string "" to unset.'
+    ),
+)
+async def put_transcription_provider(
+    body: dict = Body(
+        ...,
+        description=(
+            'Provider ID, e.g. {"provider_id": "openai"} '
+            'or {"provider_id": ""} to unset'
+        ),
+    ),
+) -> dict:
+    """Set the transcription provider."""
+    provider_id = (body.get("provider_id") or "").strip()
+    config = load_config()
+    config.agents.transcription_provider_id = provider_id
+    save_config(config)
+    return {"provider_id": provider_id}
+
+
+@router.get(
+    "/running-config",
+    response_model=AgentsRunningConfig,
+    summary="Get agent running config",
+    description="Get running configuration for active agent",
+)
+async def get_agents_running_config(
+    request: Request,
+) -> AgentsRunningConfig:
+    """Get agent running configuration."""
+    workspace = await get_agent_for_request(request)
+    agent_config = load_agent_config(workspace.agent_id)
+    return agent_config.running or AgentsRunningConfig()
+
+
+@router.put(
+    "/running-config",
+    response_model=AgentsRunningConfig,
+    summary="Update agent running config",
+    description="Update running configuration for active agent",
+)
+async def put_agents_running_config(
+    running_config: AgentsRunningConfig = Body(
+        ...,
+        description="Updated agent running configuration",
+    ),
+    request: Request = None,
+) -> AgentsRunningConfig:
+    """Update agent running configuration."""
+    workspace = await get_agent_for_request(request)
+    agent_config = load_agent_config(workspace.agent_id)
+    agent_config.running = running_config
+    save_agent_config(workspace.agent_id, agent_config)
+
+    schedule_agent_reload(request, workspace.agent_id)
+
+    return running_config
+
+
+@router.get(
+    "/system-prompt-files",
+    response_model=list[str],
+    summary="Get system prompt files",
+    description="Get system prompt files for active agent",
+)
+async def get_system_prompt_files(
+    request: Request,
+) -> list[str]:
+    """Get list of enabled system prompt files."""
+    workspace = await get_agent_for_request(request)
+    agent_config = load_agent_config(workspace.agent_id)
+    return agent_config.system_prompt_files or []
+
+
+@router.put(
+    "/system-prompt-files",
+    response_model=list[str],
+    summary="Update system prompt files",
+    description="Update system prompt files for active agent",
+)
+async def put_system_prompt_files(
+    files: list[str] = Body(
+        ...,
+        description="Markdown filenames to load into system prompt",
+    ),
+    request: Request = None,
+) -> list[str]:
+    """Update list of enabled system prompt files."""
+    workspace = await get_agent_for_request(request)
+    agent_config = load_agent_config(workspace.agent_id)
+    agent_config.system_prompt_files = files
+    save_agent_config(workspace.agent_id, agent_config)
+
+    schedule_agent_reload(request, workspace.agent_id)
+
+    return files
 
 
 # ---------------------------------------------------------------------------
@@ -105,7 +607,7 @@ def _validate_and_extract_zip(data: bytes, workspace_dir: Path) -> None:
 
 
 # ---------------------------------------------------------------------------
-# Endpoints
+# Workspace Download/Upload Endpoints
 # ---------------------------------------------------------------------------
 
 
@@ -125,7 +627,6 @@ def _validate_and_extract_zip(data: bytes, workspace_dir: Path) -> None:
 )
 async def download_workspace(request: Request):
     """Stream agent workspace as a zip file."""
-    from ..agent_context import get_agent_for_request
 
     agent = await get_agent_for_request(request)
     workspace_dir = agent.workspace_dir
@@ -172,7 +673,6 @@ async def upload_workspace(
     """
     Merge uploaded zip contents into agent workspace (overwrite, not clear).
     """
-    from ..agent_context import get_agent_for_request
 
     if file.content_type and file.content_type not in (
         "application/zip",
